Split log into separate file
